blur bugfix, very naive blur
git-svn-id: https://projectm.svn.sourceforge.net/svnroot/projectm/trunk@1099 6778bc44-b910-0410-a7a0-be141de4315d

+++ src/projectM-engine/Shader.hpp
@@ -13,13 +13,13 @@
 class Shader
 {
 public:
-	int blur1n;
-	int blur2n;
-	int blur3n;
-	int blur1x;
-	int blur2x;
-	int blur3x;
-	int blur1ed;
+	float blur1n;
+	float blur2n;
+	float blur3n;
+	float blur1x;
+	float blur2x;
+	float blur3x;
+	float blur1ed;
 
 	std::string warp;
 	std::string composite;
